BAH-4812 | Merge FhirToObservationTransformer into FhirObservationTransformer
