+++ examples/react19-consumer-app/src/App.jsx
@@ -1,15 +1,18 @@
 import React, { useRef, useState, useCallback, useEffect } from 'react';
-import { CarbonContainer } from '@bahmni/form2-controls';
+import { CarbonContainer, getObservationsFromFhir } from '@bahmni/form2-controls';
 import '@bahmni/form2-controls/dist/bundle.css';
 import '@bahmni/design-system/styles';
 import {
   clinicalFormMetadata,
-  existingObservations,
   patientData,
   translationData
 } from './formData';
 import './App.css';
-import formData from './form.json';
+import sampleFhirBundle from './sampleFhirBundle.json';
+
+// BAH-4812 demo: reverse a real FHIR searchset Bundle (the shape the backend
+// $fetch-all returns) into Form2 observations, then render them in the form.
+const existingObservations = getObservationsFromFhir(sampleFhirBundle);
 
 /**
  * React 19 Demo Application - Clinical Assessment Form
@@ -285,7 +288,7 @@ function App() {
               <ErrorBoundary>
                 <CarbonContainer
                   ref={formRef}
-                  metadata={formData}
+                  metadata={clinicalFormMetadata}
                   observations={observations}
                   patient={patientData}
                   translations={translationData}
