+++ src/helpers/FhirObservationTransformer.js
@@ -2,14 +2,313 @@ import {
   FHIR_OBSERVATION_INTERPRETATION_SYSTEM,
   FHIR_OBSERVATION_FORM_NAMESPACE_PATH_URL,
   FHIR_OBSERVATION_COMPLEX_DATA_URL,
+  FHIR_OBSERVATION_VALUE_ATTACHMENT_URL,
   CONCEPT_DATATYPE_NUMERIC,
   CONCEPT_DATATYPE_COMPLEX,
   FHIR_OBSERVATION_STATUS_FINAL,
   FHIR_RESOURCE_TYPE_OBSERVATION,
   DATE_REGEX_PATTERN,
   INTERPRETATION_TO_CODE,
+  CODE_TO_INTERPRETATION,
 } from 'src/constants/fhir';
 import { NUMBER, STRING, BOOLEAN, OBJECT } from 'src/constants';
+
+// ============================================================================
+// Reverse Transformation: FHIR Observation → Form2 Observation
+// ============================================================================
+
+const isObject = (value) => typeof value === 'object' && value !== null;
+
+const normaliseInput = (input) => {
+  if (!input) return [];
+
+  if (!Array.isArray(input) && input.resourceType === 'Bundle') {
+    return Array.isArray(input.entry) ? input.entry.filter(isObject) : [];
+  }
+
+  if (!Array.isArray(input)) return [];
+  if (input.length === 0) return [];
+
+  const first = input.find(isObject);
+  if (first && ('fullUrl' in first || 'resource' in first)) {
+    return input.filter(isObject);
+  }
+
+  return input
+    .filter(isObject)
+    .map((res) => ({ resource: res, fullUrl: `Observation/${res.id || ''}` }));
+};
+
+const buildResourceIndex = (entries) => {
+  const index = new Map();
+  for (const entry of entries) {
+    const { resource, fullUrl } = entry;
+    if (!resource) continue;
+
+    if (fullUrl) {
+      index.set(fullUrl, resource);
+    }
+    if (resource.id) {
+      index.set(`urn:uuid:${resource.id}`, resource);
+      index.set(`Observation/${resource.id}`, resource);
+      index.set(resource.id, resource);
+    }
+  }
+  return index;
+};
+
+const collectChildRefs = (entries) => {
+  const childRefs = new Set();
+  for (const entry of entries) {
+    const { resource } = entry;
+    if (!resource || !Array.isArray(resource.hasMember)) continue;
+    for (const ref of resource.hasMember) {
+      if (ref && ref.reference) {
+        childRefs.add(ref.reference);
+      }
+    }
+  }
+  return childRefs;
+};
+
+const resolveReference = (reference, index) => {
+  if (!reference) return undefined;
+  return (
+    index.get(reference) ||
+    index.get(`urn:uuid:${reference}`) ||
+    index.get(`Observation/${reference}`) ||
+    undefined
+  );
+};
+
+const findAttachmentExtension = (resource) => {
+  if (!Array.isArray(resource.extension)) return undefined;
+  return resource.extension.find(
+    (ext) =>
+      ext &&
+      ext.valueAttachment &&
+      (ext.url === FHIR_OBSERVATION_VALUE_ATTACHMENT_URL || ext.url === undefined)
+  );
+};
+
+const inferDatatype = (resource) => {
+  if (resource.valueQuantity !== undefined) return 'Numeric';
+  if (resource.valueBoolean !== undefined) return 'Boolean';
+  if (resource.valueDateTime !== undefined) return 'Date';
+  if (resource.valueCodeableConcept !== undefined) return 'Coded';
+  if (resource.valueAttachment !== undefined) return 'Complex';
+  if (findAttachmentExtension(resource)) return 'Complex';
+  return 'Text';
+};
+
+const extractValue = (resource) => {
+  if (resource.valueQuantity !== undefined) {
+    return resource.valueQuantity.value;
+  }
+
+  if (resource.valueBoolean !== undefined) {
+    return resource.valueBoolean;
+  }
+
+  if (resource.valueDateTime !== undefined) {
+    return resource.valueDateTime;
+  }
+
+  if (resource.valueCodeableConcept !== undefined) {
+    const coding =
+      resource.valueCodeableConcept.coding &&
+      resource.valueCodeableConcept.coding[0];
+    if (coding) {
+      return { uuid: coding.code, display: coding.display };
+    }
+    return undefined;
+  }
+
+  if (resource.valueAttachment !== undefined) {
+    const att = resource.valueAttachment;
+    return {
+      url: att.url,
+      fileName: att.title,
+      contentType: att.contentType,
+    };
+  }
+
+  const attachmentExt = findAttachmentExtension(resource);
+  if (attachmentExt) {
+    const att = attachmentExt.valueAttachment;
+    return {
+      url: att.url,
+      fileName: att.title,
+      contentType: att.contentType,
+    };
+  }
+
+  if (resource.valueString !== undefined) {
+    return resource.valueString;
+  }
+
+  return undefined;
+};
+
+const mapObservation = (resource, resourceIndex) => {
+  const coding = resource.code && resource.code.coding && resource.code.coding[0];
+  const conceptUuid = coding ? coding.code : undefined;
+
+  const conceptDisplay =
+    (resource.code && resource.code.text) ||
+    (coding && coding.display) ||
+    undefined;
+
+  const datatype = inferDatatype(resource);
+
+  const concept = { uuid: conceptUuid, datatype };
+  if (conceptDisplay) {
+    concept.display = conceptDisplay;
+  }
+
+  const obs = { concept };
+
+  if (resource.effectiveDateTime) {
+    obs.obsDatetime = resource.effectiveDateTime;
+  }
+
+  if (Array.isArray(resource.hasMember) && resource.hasMember.length > 0) {
+    obs.value = null;
+    const groupMembers = [];
+    for (const ref of resource.hasMember) {
+      const childResource = resolveReference(ref.reference, resourceIndex);
+      if (!childResource) {
+        console.warn(
+          'FhirObservationTransformer: Could not resolve hasMember reference',
+          ref.reference
+        );
+        continue;
+      }
+      try {
+        const childObs = mapObservation(childResource, resourceIndex);
+        if (childObs) {
+          groupMembers.push(childObs);
+        }
+      } catch (err) {
+        console.warn(
+          'FhirObservationTransformer: Error mapping hasMember child',
+          ref.reference,
+          err
+        );
+      }
+    }
+    if (groupMembers.length > 0) {
+      obs.groupMembers = groupMembers;
+    }
+  } else {
+    obs.value = extractValue(resource);
+    if (obs.value === undefined) {
+      obs.value = null;
+    }
+  }
+
+  if (Array.isArray(resource.extension)) {
+    const formPathExt = resource.extension.find(
+      (ext) => ext.url === FHIR_OBSERVATION_FORM_NAMESPACE_PATH_URL
+    );
+    if (formPathExt && formPathExt.valueString) {
+      const caretIdx = formPathExt.valueString.indexOf('^');
+      if (caretIdx !== -1) {
+        obs.formNamespace = formPathExt.valueString.slice(0, caretIdx);
+        obs.formFieldPath = formPathExt.valueString.slice(caretIdx + 1);
+      }
+    }
+  }
+
+  if (Array.isArray(resource.note)) {
+    const noteText = resource.note
+      .map((note) => note && note.text)
+      .filter(Boolean)
+      .join('\n');
+    if (noteText) {
+      obs.comment = noteText;
+    }
+  }
+
+  if (Array.isArray(resource.interpretation) && resource.interpretation.length > 0) {
+    const interpCoding =
+      resource.interpretation[0].coding && resource.interpretation[0].coding[0];
+    if (interpCoding && interpCoding.code) {
+      const word = CODE_TO_INTERPRETATION[interpCoding.code];
+      if (word) {
+        obs.interpretation = word;
+      }
+    }
+  }
+
+  return obs;
+};
+
+/**
+ * Transform a FHIR Observation Bundle (or array) back into plain form2 observation objects.
+ * This is the exact inverse of `getFhirObservations`.
+ *
+ * Accepts:
+ *   - A FHIR Bundle { resourceType: 'Bundle', entry: [{resource, fullUrl}] }
+ *   - An array of bundle entries [{resource, fullUrl}]
+ *   - An array of raw Observation resources
+ *
+ * @param {Object|Array|null|undefined} input - FHIR data to transform
+ * @returns {Array} Array of plain form2 observation objects
+ */
+export function getObservationsFromFhir(input) {
+  if (input === null || input === undefined) return [];
+
+  let entries;
+  try {
+    entries = normaliseInput(input);
+  } catch (err) {
+    console.warn('FhirObservationTransformer: Failed to normalise input', err);
+    return [];
+  }
+
+  if (!entries || entries.length === 0) return [];
+
+  const resourceIndex = buildResourceIndex(entries);
+  const childRefs = collectChildRefs(entries);
+
+  const isTopLevel = (entry) => {
+    const { fullUrl, resource } = entry;
+    if (fullUrl && childRefs.has(fullUrl)) return false;
+    if (resource && resource.id) {
+      if (childRefs.has(`urn:uuid:${resource.id}`)) return false;
+      if (childRefs.has(`Observation/${resource.id}`)) return false;
+      if (childRefs.has(resource.id)) return false;
+    }
+    return true;
+  };
+
+  const topLevelEntries = entries.filter(isTopLevel);
+
+  const results = [];
+  for (const entry of topLevelEntries) {
+    const { resource } = entry;
+    if (!resource) continue;
+    try {
+      const obs = mapObservation(resource, resourceIndex);
+      if (obs) {
+        results.push(obs);
+      }
+    } catch (err) {
+      console.warn(
+        'FhirObservationTransformer: Error mapping observation, skipping',
+        resource && resource.id,
+        err
+      );
+    }
+  }
+
+  return results;
+}
+
+// ============================================================================
+// Forward Transformation: Form2 Observation → FHIR Observation
+// ============================================================================
 
 const createCoding = (code, systemURL, display) => {
   const coding = { code };
