+++ src/index.jsx
@@ -66,8 +66,7 @@ export { CodedValueMapper } from 'src/mapper/CodedValueMapper';
 export { CodedMultiSelectValueMapper } from 'src/mapper/CodedMultiSelectValueMapper';
 
 // -------------------------- FHIR transformation ---------------------
-export { getFhirObservations } from 'src/helpers/FhirObservationTransformer';
-export { getObservationsFromFhir } from 'src/helpers/FhirToObservationTransformer';
+export { getFhirObservations, getObservationsFromFhir } from 'src/helpers/FhirObservationTransformer';
 import * as FhirConstants from 'src/constants/fhir';
 export { FhirConstants };
